에러 핸들링, 에러 페이지 만듦

+++ src/main/java/zunik/ohimarket/service/CommentService.java
@@ -8,10 +8,12 @@ import zunik.ohimarket.controller.dto.CommentUpdateDto;
 import zunik.ohimarket.domain.Comment;
 import zunik.ohimarket.domain.Member;
 import zunik.ohimarket.domain.Post;
+import zunik.ohimarket.exception.AccessDeniedException;
+import zunik.ohimarket.exception.CommentNotFoundException;
+import zunik.ohimarket.exception.PostNotFoundException;
 import zunik.ohimarket.repository.CommentRepository;
 import zunik.ohimarket.repository.PostRepository;
 
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -21,11 +23,12 @@ public class CommentService {
     private final CommentRepository commentRepository;
 
     @Transactional
-    public void save(CommentCreateDto createParam, Member loginMember) {
+    public void save(CommentCreateDto createParam, Member loginMember) throws PostNotFoundException {
         String postToken = createParam.getPostToken();
 
-        // TODO 상위 Post가 삭제되지 않았는지 확인해야함
-        Post post = postRepository.findByToken(postToken).get();
+        Post post = postRepository.findByToken(postToken).orElseThrow(
+                () -> new PostNotFoundException()
+        );
 
         Comment comment = new Comment();
         comment.setPostId(post.getId());
@@ -38,27 +41,48 @@ public class CommentService {
     }
 
     @Transactional(readOnly = true)
-    public Optional<Comment> findByToken(String commentToken) {
-        return commentRepository.findByToken(commentToken);
+    public Comment findByToken(String commentToken) throws CommentNotFoundException {
+        // 이미 댓글이 지워졌을 수 있음
+        return commentRepository.findByToken(commentToken).orElseThrow(
+                () -> new CommentNotFoundException()
+        );
     }
 
     @Transactional
-    public String update(String commentToken, CommentUpdateDto updateParam) {
-        Comment comment = commentRepository.findByToken(commentToken).get();
-        Post post = postRepository.findById(comment.getPostId()).get();
+    public String update(String commentToken, CommentUpdateDto updateParam, Long memberId) throws PostNotFoundException, CommentNotFoundException {
+        // 이미 댓글이 지워졌을 수 있습니다.
+        Comment comment = commentRepository.findByToken(commentToken).orElseThrow(
+                () -> new CommentNotFoundException()
+        );
+        // 이미 게시글이 지워졌을 수 있습니다.
+        Post post = postRepository.findById(comment.getPostId()).orElseThrow(
+                () -> new PostNotFoundException()
+        );
 
-        // TODO 인증자의 댓글이 아닐경우 삭제 실패 (memberId) 가 아니라 찾아야함
+        if (comment.getMember().getId() != memberId) {
+            // 수정하려는 댓글에 권한이 없음
+            throw new AccessDeniedException();
+        }
 
         comment.setContent(updateParam.getContent());
         return post.getToken();
     }
 
     @Transactional
-    public String delete(String token, Long memberId) {
-        Comment comment = commentRepository.findByToken(token).get();
-        Post post = postRepository.findById(comment.getPostId()).get();
+    public String delete(String token, Long memberId) throws PostNotFoundException, CommentNotFoundException, AccessDeniedException  {
+        // 이미 댓글이 지워졌을 수 있습니다.
+        Comment comment = commentRepository.findByToken(token).orElseThrow(
+                () -> new CommentNotFoundException()
+        );
+        // 이미 게시글이 지워졌을 수 있습니다.
+        Post post = postRepository.findById(comment.getPostId()).orElseThrow(
+                () -> new PostNotFoundException()
+        );
 
-        // TODO 인증자의 댓글이 아닐경우 삭제 실패
+        if (comment.getMember().getId() != memberId) {
+            // 인증자의 댓글이 아닐경우 삭제 실패
+            throw new AccessDeniedException();
+        }
 
         commentRepository.delete(comment);
         postRepository.decreaseCommentCount(comment.getPostId());
